requirements.txt
Flask
opencv-python-headless
pillow
paddleocr
paddlepaddle
numpy
requests